requirements.txt
# Core packages
llama-index==0.12.12
python-dotenv==1.0.1
ipython==8.31.0
nest-asyncio==1.6.0

# Embedding packages
llama-index-embeddings-huggingface==0.5.1
llama-index-embeddings-instructor==0.3.0
sentence-transformers==2.7.0
instructorembedding==1.0.1

# Web interface
streamlit==1.41.1

# Jupyter environment
jupyter==1.1.1
ipywidgets==8.1.5
jupyterlab==4.3.4
notebook==7.3.2

# Dependencies
huggingface-hub==0.27.1
torch>=2.1.2,<3.0.0
numpy>=1.22.0
pandas>=1.3.0
altair==5.5.0 